Move toggle to drawer
