+++ apps/parsley/src/analytics/preferences/usePreferencesAnalytics.ts
@@ -18,7 +18,6 @@ type Action =
   | { name: "Toggled zebra stripes"; on: boolean }
   | { name: "Toggled jump to failing line"; on: boolean }
   | { name: "Toggled highlight filters"; on: boolean }
-  | { name: "Toggled hide filters"; on: boolean }
   | { name: "Toggled sections"; on: boolean };
 
 export const usePreferencesAnalytics = () =>
