+++ apps/parsley/src/components/DetailsMenu/DetailsMenuCard/Toggles/index.tsx
@@ -2,7 +2,6 @@ import BaseToggle from "./BaseToggle";
 import CaseSensitiveToggle from "./CaseSensitiveToggle";
 import ExpandableRowsToggle from "./ExpandableRowsToggle";
 import FilterLogicToggle from "./FilterLogicToggle";
-import HideFiltersToggle from "./HideFilters";
 import HighlightFiltersToggle from "./HighlightFiltersToggle";
 import JumpToFailingLineToggle from "./JumpToFailingLineToggle";
 import PrettyPrintToggle from "./PrettyPrintToggle";
@@ -23,5 +22,4 @@ export {
   WordWrapFormatToggle,
   WrapToggle,
   ZebraStripingToggle,
-  HideFiltersToggle,
 };
