+++ apps/parsley/src/components/DetailsMenu/DetailsMenuCard/index.tsx
@@ -10,7 +10,6 @@ import {
   CaseSensitiveToggle,
   ExpandableRowsToggle,
   FilterLogicToggle,
-  HideFiltersToggle,
   HighlightFiltersToggle,
   JumpToFailingLineToggle,
   PrettyPrintToggle,
@@ -47,7 +46,6 @@ const DetailsMenuCard = forwardRef<HTMLDivElement, DetailsMenuProps>(
                 <CaseSensitiveToggle />
                 <FilterLogicToggle />
                 <HighlightFiltersToggle />
-                <HideFiltersToggle />
               </Column>
             </Row>
             <ButtonRow />
