+++ apps/parsley/src/components/SidePanel/NavGroup/FilterNavGroup/index.tsx
@@ -11,6 +11,7 @@ import { useFilterParam } from "hooks/useFilterParam";
 import { Filter } from "types/logs";
 import BaseNavGroup from "../BaseNavGroup";
 import FilterGroup from "./FilterGroup";
+import ShowFiltersToggle from "./ShowFiltersToggle";
 
 const { green } = palette;
 
@@ -81,13 +82,16 @@ const FilterNavGroup: React.FC<FilterNavGroupProps> = ({
       <ProjectFiltersModal open={open} setOpen={setOpen} />
       <BaseNavGroup
         additionalHeaderText={
-          <ModalTrigger
-            onClick={() => setOpen(true)}
-            role="button"
-            tabIndex={0}
-          >
-            View project filters
-          </ModalTrigger>
+          <>
+            <ShowFiltersToggle />
+            <ModalTrigger
+              onClick={() => setOpen(true)}
+              role="button"
+              tabIndex={0}
+            >
+              View project filters
+            </ModalTrigger>
+          </>
         }
         data-cy="filters"
         defaultMessage="No filters have been applied."
