+++ apps/parsley/src/context/LogContext/index.tsx
@@ -142,7 +142,6 @@ const LogContextProvider: React.FC<LogContextProviderProps> = ({
   const [highlightFilters, setHighlightFilters] = useState(
     Cookie.get(HIGHLIGHT_FILTERS) === "true",
   );
-  const [hideFilters, setHideFilters] = useState(false);
 
   const { dispatch, state } = useLogState(initialLogLines);
   const [processedLogLines, setProcessedLogLines] = useState<ProcessedLogLines>(
@@ -303,7 +302,6 @@ const LogContextProvider: React.FC<LogContextProviderProps> = ({
         caseSensitive: state.searchState.caseSensitive,
         expandableRows,
         filterLogic,
-        hideFilters,
         highlightFilters,
         prettyPrint,
         setCaseSensitive: (v: boolean) => {
@@ -317,9 +315,6 @@ const LogContextProvider: React.FC<LogContextProviderProps> = ({
         setFilterLogic: (v: FilterLogic) => {
           setFilterLogic(v);
           Cookie.set(FILTER_LOGIC, v, { expires: 365 });
-        },
-        setHideFilters: (v: boolean) => {
-          setHideFilters(v);
         },
         setHighlightFilters: (v: boolean) => {
           setHighlightFilters(v);
@@ -384,7 +379,6 @@ const LogContextProvider: React.FC<LogContextProviderProps> = ({
     [
       expandableRows,
       filterLogic,
-      hideFilters,
       lowerRange,
       matchingLines,
       prettyPrint,
@@ -411,7 +405,6 @@ const LogContextProvider: React.FC<LogContextProviderProps> = ({
       setFilterLogic,
       setLogMetadata,
       setHighlightFilters,
-      setHideFilters,
       highlightFilters,
       sectioning,
       openSectionAndScrollToLine,
