+++ apps/parsley/src/context/LogContext/types.ts
@@ -37,7 +37,6 @@ interface Preferences {
   wrap: boolean;
   zebraStriping: boolean;
   highlightFilters: boolean;
-  hideFilters: boolean;
   setCaseSensitive: (caseSensitive: boolean) => void;
   setExpandableRows: (expandableRows: boolean) => void;
   setFilterLogic: (filterLogic: FilterLogic) => void;
@@ -46,7 +45,6 @@ interface Preferences {
   setPrettyPrint: (prettyPrint: boolean) => void;
   setZebraStriping: (zebraStriping: boolean) => void;
   setHighlightFilters: (highlightFilters: boolean) => void;
-  setHideFilters: (hideFilters: boolean) => void;
 }
 
 enum DIRECTION {
